Moved cli to applications
