+++ gpt_engineer/applications/cli/main.py
@@ -36,8 +36,8 @@ from dotenv import load_dotenv
 from gpt_engineer.data.file_repository import FileRepository, FileRepositories, archive
 from gpt_engineer.core.ai import AI
 from gpt_engineer.core.steps import STEPS, Config as StepsConfig
-from gpt_engineer.cli.collect import collect_learnings
-from gpt_engineer.cli.learning import check_collection_consent
+from gpt_engineer.applications.cli.collect import collect_learnings
+from gpt_engineer.applications.cli.learning import check_collection_consent
 from gpt_engineer.data.code_vector_repository import CodeVectorRepository
 
 app = typer.Typer()  # creates a CLI app
