+++ tests/test_collect.py
@@ -10,9 +10,9 @@ from unittest.mock import MagicMock
 import pytest
 import rudderstack.analytics as rudder_analytics
 
-from gpt_engineer.cli.collect import collect_learnings, steps_file_hash
+from gpt_engineer.applications.cli import collect_learnings, steps_file_hash
 from gpt_engineer.data.file_repository import FileRepository, FileRepositories
-from gpt_engineer.cli.learning import extract_learning
+from gpt_engineer.applications.cli import extract_learning
 from gpt_engineer.core.steps import simple_gen
